+++ src/AppBundle/Service/MonitoringSetService.php
@@ -8,6 +8,7 @@ use AppBundle\Document\MonitoringSet;
 use AppBundle\Manager\TemplateManager;
 use JMS\Serializer\SerializerInterface;
 use AppBundle\Manager\MonitoringSetManager;
+use AppBundle\Manager\ApplicationTypeManager;
 
 /**
  * Service class for MonitoringSet entities
@@ -19,6 +20,11 @@ class MonitoringSetService
      * @var MonitoringSetManager
      */
     private $monitoringSetManager;
+
+    /**
+     * @var ApplicationTypeManager
+     */
+    private $applicationTypeManager;
 
     /**
      * @var TemplateManager
@@ -36,20 +42,22 @@ class MonitoringSetService
     private $logger;
 
     /**
-     * Constructor
      *
      * @param MonitoringSetManager $monitoringSetManager
+     * @param ApplicationTypeManager $applicationTypeManager
      * @param TemplateManager $templateManager
      * @param SerializerInterface $serializer
      * @param LoggerInterface $logger
      */
     public function __construct(
         MonitoringSetManager $monitoringSetManager,
+        ApplicationTypeManager $applicationTypeManager,
         TemplateManager $templateManager,
         SerializerInterface $serializer,
         LoggerInterface $logger
     ) {
         $this->monitoringSetManager = $monitoringSetManager;
+        $this->applicationTypeManager = $applicationTypeManager;
         $this->templateManager = $templateManager;
         $this->serializer = $serializer;
         $this->logger = $logger;
@@ -64,6 +72,11 @@ class MonitoringSetService
     public function listMonitoringSets()
     {
         return $this->monitoringSetManager->getAll();
+    }
+
+    public function listValidMonitoringSets()
+    {
+        return $this->monitoringSetManager->getValid();
     }
 
     /**
@@ -118,7 +131,17 @@ class MonitoringSetService
      */
     public function newMonitoringSet($json)
     {
+        /** @var MonitoringSet $monitoringSet */
         $monitoringSet = $this->serializer->deserialize($json, MonitoringSet::class, 'json');
+        $templates = $monitoringSet->getTemplates();
+        $monitoringSet->resetTemplates();
+        foreach ($templates as $template) {
+            $loaded = $this->templateManager->getOneBy(['id' => $template->getId()]);
+            if (($loaded instanceof Template) === false) {
+                continue;
+            }
+            $monitoringSet->addTemplate($loaded);
+        }
         $this->monitoringSetManager->update($monitoringSet);
 
         return true;
@@ -138,8 +161,26 @@ class MonitoringSetService
         $isSuccessful = false;
 
         try {
+            /** @var MonitoringSet $monitoringSet */
             $monitoringSet = $this->serializer->deserialize($json, MonitoringSet::class, 'json');
-            $this->monitoringSetManager->update($monitoringSet);
+
+            $dbDocument = $this->monitoringSetManager->getOneBy(['id' => $monitoringSet->getId()]);
+
+            if ($dbDocument instanceof MonitoringSet) {
+                $dbDocument->setVersion($monitoringSet->getVersion());
+                $dbDocument->setQualifier($monitoringSet->getQualifier());
+                $dbDocument->setName($monitoringSet->getName());
+                $dbDocument->setVersion($monitoringSet->getVersion());
+                $dbDocument->resetTemplates();
+                foreach ($monitoringSet->getTemplates() as $template) {
+                    $loaded = $this->templateManager->getOneBy(['id' => $template->getId()]);
+                    if (($loaded instanceof Template) === false) {
+                        continue;
+                    }
+                    $dbDocument->addTemplate($loaded);
+                }
+                $this->monitoringSetManager->update($dbDocument);
+            }
             $isSuccessful = true;
         } catch (\Exception $e) {
             $this->logger->error($e->getMessage());
@@ -161,13 +202,13 @@ class MonitoringSetService
     {
         $ms = $this->monitoringSetManager->getOneBy(['id' => $id]);
 
-        if ($ms instanceof MonitoringSet) {
-            /** @var Template $template */
-            foreach ($ms->getTemplates() as $template) {
-                $template->setMonitoringSet(null);
-                $this->templateManager->update($template);
-            }
+        $linkedTypes = $this->applicationTypeManager->getLinkedTypes($ms);
 
+        if (count($linkedTypes) > 0) {
+            throw new \Exception("Cannot delete a monitoring set that is used in an application type");
+        }
+
+        if ($ms instanceof MonitoringSet) {
             $this->monitoringSetManager->deleteById($id);
         }
     }
